refactor: package 구조 세부화
