--- app/src/main/java/woowacourse/omok/domain/Users.kt
@@ -1,3 +1,0 @@
-package woowacourse.omok.domain
-
-class Users(val value: List<User>)
