+++ app/src/main/java/woowacourse/omok/room/db/Column.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok.db
+package woowacourse.omok.room.db
 
 data class Column(
     val name: String,
