+++ app/src/main/java/woowacourse/omok/room/db/SQLiteTable.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok.db
+package woowacourse.omok.room.db
 
 interface SQLiteTable {
     val name: String
