+++ app/src/main/java/woowacourse/omok/room/db/table/UserStagesTable.kt
@@ -1,7 +1,7 @@
-package woowacourse.omok.db.table
+package woowacourse.omok.room.db.table
 
-import woowacourse.omok.db.Column
-import woowacourse.omok.db.SQLiteTable
+import woowacourse.omok.room.db.Column
+import woowacourse.omok.room.db.SQLiteTable
 
 object UserStagesTable : SQLiteTable {
     const val USER_ID = "userId"
