+++ app/src/main/java/woowacourse/omok/room/domain/Stage.kt
@@ -1,3 +1,3 @@
-package woowacourse.omok.domain
+package woowacourse.omok.room.domain
 
 class Stage(val id: Int, val value: StageStones = StageStones(listOf()))
