+++ app/src/main/java/woowacourse/omok/room/domain/Stages.kt
@@ -1,3 +1,3 @@
-package woowacourse.omok.domain
+package woowacourse.omok.room.domain
 
 class Stages(val value: List<Stage>)
