+++ app/src/main/java/woowacourse/omok/room/domain/User.kt
@@ -1,3 +1,3 @@
-package woowacourse.omok.domain
+package woowacourse.omok.room.domain
 
 data class User(val id: Int, val name: String)
