+++ app/src/main/java/woowacourse/omok/room/domain/Users.kt
@@ -1,0 +1,3 @@
+package woowacourse.omok.room.domain
+
+class Users(val value: List<User>)
