+++ app/src/main/java/woowacourse/omok/room/view/RoomEventObservable.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok.view
+package woowacourse.omok.room.view
 
 fun interface RoomEventObservable {
     fun event()
