+++ app/src/main/java/woowacourse/omok/room/view/RoomEventObserver.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok.view
+package woowacourse.omok.room.view
 
 class RoomEventObserver {
     private val roomEvent: MutableList<RoomEventObservable> = mutableListOf()
